udpate

+++ classification/models/TextCNN.py
@@ -20,7 +20,7 @@ class TextCNN(object):
         self.input_x = tf.placeholder(dtype=tf.int32, shape=[None, self.seq_len], name='input_x')
         self.input_y = tf.placeholder(dtype=tf.float32, shape=[None, self.class_num], name='input_y')
         self.global_step = tf.placeholder(shape=(), dtype=tf.int32, name='global_step')
-        self.keep_prob = tf.placeholder(tf.float32, name='keep_prob')
+        self.keep_prob = tf.placeholder(tf.float64, name='keep_prob')
 
         self.embedding_matrix = tokenizer.embedding_matrix
         self.cnn()
